final moving of functions
